backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
google-genai
replicate
python-dotenv
httpx
